Added MS effects for sensor traversed by the helix hypothesis but without a hit associated to the track
